--- a/app/controllers/system/admin/courses_controller.rb
+++ b/app/controllers/system/admin/courses_controller.rb
@@ -38,7 +38,6 @@ class System::Admin::CoursesController < System::Admin::Controller
 
     @courses = @courses.ordered_by_title
     @courses_count = @courses.count.is_a?(Hash) ? @courses.count.count : @courses.count
-    @courses = @courses.paginated(new_page_params)
 
     @owner_preload_service = Course::CourseOwnerPreloadService.new(@courses.map(&:id))
   end