--- a/client/app/bundles/course/assessment/pages/AssessmentsIndex/index.tsx
+++ b/client/app/bundles/course/assessment/pages/AssessmentsIndex/index.tsx
@@ -73,12 +73,7 @@ const AssessmentsIndex = (): JSX.Element => {
             </Tabs>
           )}
 
-          {refreshable(
-            <AssessmentsTable
-              assessments={data}
-              top={data.display.category.tabs.length > 1 ? 'top-40' : 'top-20'}
-            />,
-          )}
+          {refreshable(<AssessmentsTable assessments={data} />)}
         </>
       )}
     </Preload>