--- a/client/app/bundles/course/assessment/question/programming/components/common/TestCasesTable.tsx
+++ b/client/app/bundles/course/assessment/question/programming/components/common/TestCasesTable.tsx
@@ -10,7 +10,7 @@ import {
 } from '@mui/material';
 
 import Accordion from 'lib/components/core/layouts/Accordion';
-import { TableContainer } from 'lib/components/core/table';
+import TableContainer from 'lib/components/core/layouts/TableContainer';
 import useTranslation from 'lib/hooks/useTranslation';
 
 import translations from '../../../../translations';