--- a/client/app/lib/components/table/MuiTableAdapter/index.ts
+++ b/client/app/lib/components/table/MuiTableAdapter/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MuiTable';