--- a/client/app/lib/components/table/TanStackTableBuilder/index.ts
+++ b/client/app/lib/components/table/TanStackTableBuilder/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useTanStackTableBuilder';