… client/app/bundles/course/users/components/tables/ManageUsersTable.tsx
… client/app/bundles/course/users/components/tables/ManageUsersTable/RoleMenu.tsx
… client/app/bundles/course/users/components/tables/ManageUsersTable/TimelineMenu.tsx
… client/app/bundles/course/users/components/tables/ManageUsersTable/UserNameField.tsx
… client/app/bundles/course/users/components/tables/ManageUsersTable/index.tsx
… client/app/bundles/course/users/components/tables/ManageUsersTable/translations.ts
… client/app/bundles/course/users/pages/ManageStaff/index.tsx
… client/app/bundles/course/users/pages/ManageStudents/index.tsx
… client/app/bundles/course/video/pages/VideoShow/VideoDetails.tsx
… client/app/bundles/system/admin/admin/components/tables/CoursesTable.tsx
… client/app/bundles/system/admin/admin/components/tables/InstancesTable.tsx
… client/app/bundles/system/admin/admin/components/tables/InstancesTable/InstanceField.tsx
… client/app/bundles/system/admin/admin/components/tables/InstancesTable/index.tsx
… client/app/bundles/system/admin/admin/components/tables/ManageInstanceTable/InstanceNameOrHostField.tsx
… client/app/bundles/system/admin/admin/operations.ts
… client/app/bundles/system/admin/admin/pages/CoursesIndex/index.tsx
… client/app/bundles/system/admin/admin/pages/InstancesIndex/index.tsx
… client/app/bundles/system/admin/instance/instance/components/tables/UsersTable.tsx
… client/app/bundles/system/admin/instance/instance/pages/InstanceCoursesIndex/index.tsx
… client/app/bundles/system/admin/instance/instance/pages/InstanceUsersIndex/index.tsx
… client/app/lib/components/core/fields/SearchField.tsx
… client/app/lib/components/core/layouts/DataTable.jsx
… client/app/lib/components/core/layouts/TableContainer.tsx
… client/app/lib/components/core/table/Cell.tsx
… client/app/lib/components/core/table/HorizontalTable.tsx
… client/app/lib/components/core/table/Row.tsx
… client/app/lib/components/core/table/TableContainer.tsx
… client/app/lib/components/core/table/VerticalTable.tsx
… client/app/lib/components/core/table/index.tsx
… client/app/lib/components/core/table/template.ts