Closes #5693. Builds on the research done in #5425.

![new-tables (1)](https://user-images.githubusercontent.com/51525686/218937440-063c203f-ed76-4dfa-a77f-abfaa20833d5.svg)

| Before | After |
| - | - |
| ![image](https://user-images.githubusercontent.com/51525686/218513622-c2156f46-8b25-4e4a-9e22-566896b81c2a.png) | ![image](https://user-images.githubusercontent.com/51525686/218513484-61921ce9-1755-4e89-81a9-858c26bd6cd0.png)


## Performance
- MacBook Pro 14-inch, M1 Max (10-core CPU), 64 GB unified memory
- macOS Ventura 13.2, Chrome 110.0.5481.77 (arm64), React Developer Tools 4.27.1
- Loading 100 students in `ManageStudents`
- Re-renders were triggered by flipping the Phantom switch on the first student.
- Fresh loads were repeated 5 times and averaged.
- Re-renders comprised of 3 peaks, repeated 3 times, summed and averaged.

| | Previous | New | Summary |
| - | - | - | - |
| Fresh load | 555.44 ms | 445.6 ms | 19.8% faster |
| Re-renders | 1180.50 ms | 20.9 ms | 98.2% faster |