@purfectliterature 2023-02-20
@bivanalhar You may want to rebase #5749 and your branch [bivan.ah/forum-reorder](https://github.com/Coursemology/coursemology2/tree/bivan.ah/forum-reorder) with the latest changes here.