@purfectliterature 2023-02-20
@bivanalhar The changes should only concern the `<Table ... />` props, and all of them are just 1-1 maps from the previous props' values, though I suspect you might get a merge conflict. You may me know if you need to clarify some changes.